поддержали суффиксы у чисел, перевели все числа во внутреннем представлении c int/float на long/double.
git-svn-id: file:///home/cppg/dev/svn/compilers2/trunk@555 a9b74587-2ee6-7842-8a07-03ad48b783cd

+++ cc4x86/src/src/x86_bincode.c
@@ -266,7 +266,7 @@ void bincode_create_operand_from_label(x86_operand *op, int label)
 //  Handling floating-point constants.
 //
 
-x86_instruction_code bincode_encode_float_constant(float val)
+x86_instruction_code bincode_encode_float_constant(double val)
 {
     if (val == 0.0) {
         return x86instr_float_zero;
